Set Up for JS not working with dist, so collapsed js into main, styling tweaks, added a loader so that you have to wait until your location is given before you can query
