+++ app/styles/app.css
@@ -25,12 +25,20 @@ ul {
 }
 
 /* line 19, scss/_base.scss */
+button {
+  border: none;
+  border-radius: 3px;
+  background-color: #007dba;
+  color: #FFF;
+}
+
+/* line 28, scss/_base.scss */
 .backfaceVisibility {
   -webkit-backface-visibility: visible;
   backface-visibility: visible;
 }
 
-/* line 23, scss/_base.scss */
+/* line 32, scss/_base.scss */
 .vert-panel {
   min-height: 100%;
   display: -webkit-flex;
@@ -41,8 +49,8 @@ ul {
   align-content: stretch;
 }
 
-/* line 29, scss/_base.scss */
-.horz-panel {
+/* line 38, scss/_base.scss */
+.horz-panel, header {
   min-width: 100%;
   display: -webkit-flex;
   display: flex;
@@ -52,17 +60,55 @@ ul {
   align-content: stretch;
 }
 
-/* line 35, scss/_base.scss */
+/* line 44, scss/_base.scss */
 .overlay {
   position: absolute;
   top: 0;
   left: 0;
   width: 100%;
 }
+/* line 49, scss/_base.scss */
+.overlay.modal {
+  background-color: rgba(0, 0, 0, 0.4);
+  height: 100%;
+  z-index: 9;
+}
 
-/* line 41, scss/_base.scss */
+/* line 55, scss/_base.scss */
 .name {
   font-weight: 600;
+}
+
+/* line 58, scss/_base.scss */
+#directions {
+  background: #fff;
+  padding: .5em;
+  position: absolute;
+  top: 0;
+  left: 0;
+}
+/* line 64, scss/_base.scss */
+#directions #directions-map {
+  width: 100%;
+  height: 450px;
+}
+
+/* line 69, scss/_base.scss */
+.loading {
+  display: -webkit-flex;
+  display: flex;
+  -webkit-flex-direction: column;
+  flex-direction: column;
+  -webkit-align-content: center;
+  align-content: center;
+  -webkit-justify-content: center;
+  justify-content: center;
+  text-align: center;
+}
+/* line 75, scss/_base.scss */
+.loading * {
+  -webkit-align-self: center;
+  align-self: center;
 }
 
 /* line 1, scss/_map.scss */
@@ -92,17 +138,13 @@ header {
 header * {
   margin: .25em .5em;
 }
-/* line 15, scss/_searchbar.scss */
-header h1 {
-  font-size: .7em;
-}
-/* line 17, scss/_searchbar.scss */
+/* line 16, scss/_searchbar.scss */
 header button {
   font-size: 1.5em;
   padding: 0 1em;
 }
 
-/* line 22, scss/_searchbar.scss */
+/* line 21, scss/_searchbar.scss */
 #searchbox {
   flex: 1;
   font-size: 2em;
@@ -111,19 +153,21 @@ header button {
 /* line 1, scss/_list.scss */
 #list-panel {
   flex: 1;
-  background-color: rgba(255, 255, 255, 0.8);
+  background-color: rgba(0, 0, 0, 0.5);
   overflow: scroll;
   padding: .5em;
 }
 
 /* line 8, scss/_list.scss */
 li {
-  border-bottom: solid 1px #DEDEDE;
+  background-color: #FFF;
+  border-bottom: solid 3px #DEDEDE;
+  border-top: solid 2px #FFF;
+  margin: .75em;
   padding: .5em;
   cursor: default;
 }
-/* line 12, scss/_list.scss */
+/* line 15, scss/_list.scss */
 li:hover {
-  background-color: #333;
-  color: #FFF;
+  border-bottom: solid 3px #AAA;
 }
