+++ app/styles/scss/_base.scss
@@ -16,6 +16,15 @@ ul {
 	margin:0;
 	padding:0;
 }
+button {
+	border:none;
+	border-radius:3px;
+	background-color:#007dba;
+	color:#FFF;
+}
+header {
+	@extend .horz-panel;
+}
 .backfaceVisibility {
   -webkit-backface-visibility: visible;
   backface-visibility: visible;
@@ -37,7 +46,33 @@ ul {
 	top:0;
 	left:0;
 	width:100%;
+	&.modal {
+		background-color:rgba(0,0,0,.4);
+		height: 100%;
+		z-index:9;
+	}
 }
 .name {
 	font-weight:600;
+}
+#directions {
+	background: #fff;
+	padding:.5em;
+	position:absolute;
+	top:0;
+	left:0;
+	#directions-map {
+		width:100%;
+		height:450px;
+	}
+}
+.loading {
+    @include display-flex;
+    @include flex-direction(column);
+	@include align-content(center);
+	@include justify-content(center);
+	text-align:center;
+	* {
+		@include align-self(center);
+	}
 }
